backend/requirements.txt
Django==4.2.27
djangorestframework==3.15.2
django-cors-headers==4.4.0
channels==4.1.0
daphne==4.1.2


